Reorganized file structure and moved queries to separate files
