+++ db/modules/index.js
@@ -1,0 +1,7 @@
+const getQuestions = require('./getQuestions');
+const getAnswers = require('./getAnswers');
+
+module.exports = {
+  getQuestions,
+  getAnswers,
+};
